Some language changes regarding enactors and survivors of violence
It should be clear that I donʼt really give a fuck about legal
definitions but for people who do work in those circles,
perpetrator/victim have specific legal connotations which do not
necessarily reflect who is the actual enactor/survivor of violence
in a given situation.  For clarity, I've changed the former to the
latter in most (although not quite all) cases.